requirements.txt
PyQt5
pygame
supabase
python-dotenv
